fix(cc-task): move non-presentational types back
